[BACKLOG-195] Removing BlockoutResource - moving other methods to util
